feat: put serve function for assets in a new file
