+++ handlers/assets/node_modules.go
@@ -32,14 +32,6 @@ var htmxWsBytes []byte = readFile("ws.min.js") // This is aliases to htmx-ws.js 
 var tailwindBytes []byte = readFile("tailwind.css")
 var faviconBytes []byte = readFile("favicon.ico")
 
-func ServeAsset(bytes []byte, contentType string) func(http.ResponseWriter, *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("content-type", contentType)
-		w.Header().Set("cache-control", "public, max-age=604800, s-max-age=604800")
-		w.Write(bytes)
-	}
-}
-
 func Register(mux *http.ServeMux, themeCss []byte) {
 	mux.HandleFunc("/assets/htmx.js", ServeAsset(htmxBytes, "application/javascript"))
 	mux.HandleFunc("/assets/htmx-ws.js", ServeAsset(htmxWsBytes, "application/javascript"))
